Prototype 1/backend/requirements.txt
fastapi
uvicorn[standard]
groq
pypdf
python-dotenv
